fixed msvc warnings

+++ src/media_tools/dash_client.c
@@ -9398,7 +9398,7 @@ GF_Err gf_dash_group_next_seg_info(GF_DashClient *dash, u32 group_idx, u32 depen
 					seg_time->den = seg_scale;
 				}
 				if (seg_dur_ms) {
-					*seg_dur_ms = gf_timestamp_rescale(segment_dur, seg_scale, 1000);
+					*seg_dur_ms = (u32) gf_timestamp_rescale(segment_dur, seg_scale, 1000);
 				}
 				group->download_segment_index = cur_grp_idx;
 			}
